reuse existent service and renamed site to library
