+++ src/app/app.module.ts
@@ -63,7 +63,6 @@ import { NodeVersionsDirective } from './common/directives/node-versions.directi
 import { NodeVersionsDialogComponent } from './dialogs/node-versions/node-versions.dialog';
 import { LibraryDialogComponent } from './dialogs/library/library.dialog';
 import { ContentManagementService } from './common/services/content-management.service';
-import { LibraryManagementService } from './common/services/library-management.service';
 import { NodeActionsService } from './common/services/node-actions.service';
 import { NodePermissionService } from './common/services/node-permission.service';
 import { SearchResultsComponent } from './components/search/search-results/search-results.component';
@@ -157,7 +156,6 @@ import { PermissionsManagerComponent } from './components/permission-manager/per
             }
         },
         ContentManagementService,
-        LibraryManagementService,
         NodeActionsService,
         NodePermissionService,
         ProfileResolver,
